feat: cleanup old GitHub packages
This commit refactors the code to include a new step in the workflow that cleans up old GitHub packages. The "package-cleanup" step is added, which runs on Ubuntu latest and deletes package versions for the gameserver container. It uses the actions/delete-package-versions@v4 action with specific parameters to delete only untagged versions and keep a minimum of 1 version.