feat: add package-cleanup workflow
This commit adds a new GitHub Actions workflow file, `package-cleanup.yml`, which is responsible for cleaning up old versions of the `gameserver` container package in GitHub Packages. The workflow runs on a schedule every Wednesday at 4:05 AM UTC.

The workflow consists of three jobs (`package-cleanup`, `package-cleanup-2`, and `package-cleanup-3`) that delete old package versions using the `actions/delete-package-versions@v4` action. The minimum number of versions to keep is set to 1, and only untagged versions are deleted.

These changes aim to automate the cleanup process and ensure that only necessary package versions are retained in GitHub Packages.